requirements-dashboard.txt
# ── Core (also in requirements.txt) ─────────────────────────
gymnasium>=0.29.1
numpy>=1.24.0
pydantic>=2.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# ── Streamlit dashboard ───────────────────────────────────────
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.0.0
pillow>=10.0.0

# ── Media analysis ────────────────────────────────────────────
opencv-python-headless>=4.8.0

# ── Text analysis ────────────────────────────────────────────
langdetect>=1.0.9
deep-translator

# ── Telegram bot ─────────────────────────────────────────────
python-telegram-bot>=20.0
